requirements.txt
click>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.31.0
PyYAML>=6.0.1 